Add main.rs to vview
